requirements.txt
yfinance
langchain-community
langchain-core
pandas
gspread
oauth2client
python-dotenv
datetime
gradio
jupyter
ipykernel
holidays
plotly
litellm
langgraph